FIX: 파일 업로드 관련 에러 수정

+++ src/main/java/com/example/airplanning/service/UserService.java
@@ -124,7 +124,9 @@ public class UserService {
             throw new AppException(ErrorCode.FILE_UPLOAD_ERROR);
         }
 
-        deleteFile(filePath);
+        if (filePath != null) {
+            deleteFile(filePath);
+        }
         return defaultUrl + fileName;
 
     }
